CLUE-568: cleanup — drop dead persisted programZoom; extract getNewIndexedName
Follow-up cleanups after the fit-on-restore change (no behavior change):
- remove the persisted programZoom prop, its setProgramZoom action, and the
  now-dead persist calls; the canvas fits content on every load, so only the
  volatile liveProgramZoom (live session transform) remains
- drop the dataflow tileSnapshotForCopy override (base returns getSnapshot(self)
  identically once the programZoom reset is gone); legacy snapshots still load,
  MST ignores the removed key
- move getNewIndexedName out of the rete controller into
  nodes/utilities/indexed-name.ts (next to its test); tidy its docstring
